+++ src/main/java/com/shotty/shotty/YoutubeService.java
@@ -1,10 +1,14 @@
 package com.shotty.shotty;
 
+import com.shotty.shotty.youtube.dto.channel.ChannelSimpleInfoDto;
+import com.shotty.shotty.youtube.dto.video.ShortsSimpleInfoDto;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
+
+import java.util.List;
 
 @Service
 public class YoutubeService {
@@ -27,6 +31,21 @@ public class YoutubeService {
         ResponseEntity<ShortsSimpleInfoDto> response = restTemplate.getForEntity(url, ShortsSimpleInfoDto.class);
 
         return response;
+    }
+
+    public ShortsSimpleInfoDto searchVideos(List<String> shortsIds){
+        RestTemplate restTemplate = new RestTemplate();
+
+        UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromHttpUrl(VIDEO_API_URL)
+                .queryParam("part", "snippet,statistics")
+                .queryParam("id", shortsIds)
+                .queryParam("key", API_KEY);
+
+        String url = uriBuilder.toUriString();
+
+        ResponseEntity<ShortsSimpleInfoDto> response = restTemplate.getForEntity(url, ShortsSimpleInfoDto.class);
+
+        return response.getBody();
     }
 
     public ResponseEntity<ChannelSimpleInfoDto> searchChannel(String channel_id){
