+++ src/main/java/com/shotty/shotty/youtube/dto/channel/ChannelSimpleInfoDto.java
@@ -1,0 +1,11 @@
+package com.shotty.shotty.youtube.dto.channel;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class ChannelSimpleInfoDto {
+    private List<YouTubeChannelItem> items;
+
+}
