+++ src/main/java/com/shotty/shotty/youtube/dto/channel/ChannelSimpleSnippet.java
@@ -1,0 +1,10 @@
+package com.shotty.shotty.youtube.dto.channel;
+
+import lombok.Data;
+
+@Data
+public class ChannelSimpleSnippet {
+    private String publishedAt;
+    private String title;
+    private String description;
+}
