+++ src/main/java/com/shotty/shotty/youtube/dto/channel/ChannelSimpleStatistics.java
@@ -1,0 +1,10 @@
+package com.shotty.shotty.youtube.dto.channel;
+
+import lombok.Data;
+
+@Data
+public class ChannelSimpleStatistics {
+    private String viewCount;
+    private String subscriberCount;
+    private String videoCount;
+}
