+++ src/main/java/com/shotty/shotty/youtube/dto/channel/YouTubeChannelItem.java
@@ -1,0 +1,10 @@
+package com.shotty.shotty.youtube.dto.channel;
+
+import lombok.Data;
+
+@Data
+public class YouTubeChannelItem {
+    private String id;
+    private ChannelSimpleSnippet snippet;
+    private ChannelSimpleStatistics statistics;
+}
