+++ src/main/java/com/shotty/shotty/youtube/dto/video/ShortsSimpleInfoDto.java
@@ -1,0 +1,13 @@
+package com.shotty.shotty.youtube.dto.video;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class ShortsSimpleInfoDto {
+    private List<YouTubeVideoItem> items;
+}
+
+
+
