+++ src/main/java/com/shotty/shotty/youtube/dto/video/VideoSimpeSnippet.java
@@ -1,0 +1,12 @@
+package com.shotty.shotty.youtube.dto.video;
+
+import lombok.Data;
+
+@Data
+public class VideoSimpeSnippet {
+    private String publishedAt;
+    private String channelId;
+    private String title;
+    private String description;
+    private String channelTitle;
+}
