+++ src/main/java/com/shotty/shotty/youtube/dto/video/VideoSimpleStatistics.java
@@ -1,0 +1,11 @@
+package com.shotty.shotty.youtube.dto.video;
+
+import lombok.Data;
+
+@Data
+public class VideoSimpleStatistics {
+    private String viewCount;
+    private String likeCount;
+    private String favoriteCount;
+    private String commentCount;
+}
