+++ src/main/java/com/shotty/shotty/youtube/dto/video/YouTubeVideoItem.java
@@ -1,0 +1,10 @@
+package com.shotty.shotty.youtube.dto.video;
+
+import lombok.Data;
+
+@Data
+public class YouTubeVideoItem {
+    private String id;
+    private VideoSimpeSnippet snippet;
+    private VideoSimpleStatistics statistics;
+}
